+++ test/assertions/index.test.ts
@@ -1,6 +1,5 @@
 import dedent from 'dedent';
 import * as fs from 'fs';
-import { createRequire } from 'node:module';
 import * as path from 'path';
 import { runAssertion, runAssertions } from '../../src/assertions';
 import { validateXml, containsXml } from '../../src/assertions/xml';
@@ -11,8 +10,6 @@ import {
   OpenAiChatCompletionProvider,
 } from '../../src/providers/openai';
 import { ReplicateModerationProvider } from '../../src/providers/replicate';
-import { runPython } from '../../src/python/pythonUtils';
-import { runPythonCode } from '../../src/python/wrapper';
 import type {
   Assertion,
   ApiProvider,
@@ -44,22 +41,6 @@ jest.mock('../../src/fetch', () => {
   return {
     ...actual,
     fetchWithRetries: jest.fn(actual.fetchWithRetries),
-  };
-});
-
-jest.mock('../../src/python/wrapper', () => {
-  const actual = jest.requireActual('../../src/python/wrapper');
-  return {
-    ...actual,
-    runPythonCode: jest.fn(actual.runPythonCode),
-  };
-});
-
-jest.mock('../../src/python/pythonUtils', () => {
-  const actual = jest.requireActual('../../src/python/pythonUtils');
-  return {
-    ...actual,
-    runPython: jest.fn(actual.runPython),
   };
 });
 
@@ -573,67 +554,6 @@ describe('runAssertion', () => {
         },
       },
     },
-  };
-
-  const javascriptStringAssertion: Assertion = {
-    type: 'javascript',
-    value: 'output === "Expected output"',
-  };
-
-  const javascriptMultilineStringAssertion: Assertion = {
-    type: 'javascript',
-    value: `
-      if (output === "Expected output") {
-        return {
-          pass: true,
-          score: 0.5,
-          reason: 'Assertion passed',
-        };
-      }
-      return {
-        pass: false,
-        score: 0,
-        reason: 'Assertion failed',
-      };`,
-  };
-
-  const javascriptStringAssertionWithNumber: Assertion = {
-    type: 'javascript',
-    value: 'output.length * 10',
-  };
-
-  const javascriptBooleanAssertionWithConfig: Assertion = {
-    type: 'javascript',
-    value: 'output.length <= context.config.maximumOutputSize',
-    config: {
-      maximumOutputSize: 20,
-    },
-  };
-
-  const javascriptStringAssertionWithNumberAndThreshold: Assertion = {
-    type: 'javascript',
-    value: 'output.length * 10',
-    threshold: 0.5,
-  };
-
-  const javascriptFunctionAssertion: Assertion = {
-    type: 'javascript',
-    value: async (output: string) => ({
-      pass: true,
-      score: 0.5,
-      reason: 'Assertion passed',
-      assertion: null,
-    }),
-  };
-
-  const javascriptFunctionFailAssertion: Assertion = {
-    type: 'javascript',
-    value: async (output: string) => ({
-      pass: false,
-      score: 0.5,
-      reason: 'Assertion failed',
-      assertion: null,
-    }),
   };
 
   it('should pass when the equality assertion passes', async () => {
@@ -1530,145 +1450,6 @@ describe('runAssertion', () => {
     );
   });
 
-  it('should pass when the javascript assertion passes', async () => {
-    const output = 'Expected output';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptStringAssertion,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: true,
-      reason: 'Assertion passed',
-    });
-  });
-
-  it('should pass a score through when the javascript returns a number', async () => {
-    const output = 'Expected output';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptStringAssertionWithNumber,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: true,
-      score: output.length * 10,
-      reason: 'Assertion passed',
-    });
-  });
-
-  it('should pass when javascript returns an output string that is smaller than the maximum size threshold', async () => {
-    const output = 'Expected output';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptBooleanAssertionWithConfig,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: true,
-      score: 1.0,
-      reason: 'Assertion passed',
-    });
-  });
-
-  it('should fail when javascript returns an output string that is larger than the maximum size threshold', async () => {
-    const output = 'Expected output with some extra characters';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptBooleanAssertionWithConfig,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: false,
-      score: 0,
-      reason: expect.stringContaining('Custom function returned false'),
-    });
-  });
-
-  it('should pass when javascript returns a number above threshold', async () => {
-    const output = 'Expected output';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptStringAssertionWithNumberAndThreshold,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: true,
-      score: output.length * 10,
-      reason: 'Assertion passed',
-    });
-  });
-
-  it('should fail when javascript returns a number below threshold', async () => {
-    const output = '';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptStringAssertionWithNumberAndThreshold,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: false,
-      score: output.length * 10,
-      reason: expect.stringContaining('Custom function returned false'),
-    });
-  });
-
-  it('should set score when javascript returns false', async () => {
-    const output = 'Test output';
-
-    const assertion: Assertion = {
-      type: 'javascript',
-      value: 'output.length < 1',
-    };
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: false,
-      score: 0,
-      reason: expect.stringContaining('Custom function returned false'),
-    });
-  });
-
-  it('should fail when the javascript assertion fails', async () => {
-    const output = 'Different output';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptStringAssertion,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: false,
-      reason: 'Custom function returned false\noutput === "Expected output"',
-    });
-  });
-
   it('should pass when assertion passes - with vars', async () => {
     const output = 'Expected output';
 
@@ -1686,113 +1467,6 @@ describe('runAssertion', () => {
     expect(result).toMatchObject({
       pass: true,
       reason: 'Assertion passed',
-    });
-  });
-
-  it('should pass when javascript function assertion passes - with vars', async () => {
-    const output = 'Expected output';
-
-    const javascriptStringAssertionWithVars: Assertion = {
-      type: 'javascript',
-      value: 'output === "Expected output" && context.vars.foo === "bar"',
-    };
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptStringAssertionWithVars,
-      test: { vars: { foo: 'bar' } } as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: true,
-      reason: 'Assertion passed',
-    });
-  });
-
-  it('should fail when the javascript does not match vars', async () => {
-    const output = 'Expected output';
-
-    const javascriptStringAssertionWithVars: Assertion = {
-      type: 'javascript',
-      value: 'output === "Expected output" && context.vars.foo === "something else"',
-    };
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptStringAssertionWithVars,
-      test: { vars: { foo: 'bar' } } as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: false,
-      reason:
-        'Custom function returned false\noutput === "Expected output" && context.vars.foo === "something else"',
-    });
-  });
-
-  it('should pass when the function returns pass', async () => {
-    const output = 'Expected output';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptFunctionAssertion,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: true,
-      score: 0.5,
-      reason: 'Assertion passed',
-    });
-  });
-
-  it('should fail when the function returns fail', async () => {
-    const output = 'Expected output';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-      assertion: javascriptFunctionFailAssertion,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-    });
-    expect(result).toMatchObject({
-      pass: false,
-      score: 0.5,
-      reason: 'Assertion failed',
-    });
-  });
-
-  it('should pass when the multiline javascript assertion passes', async () => {
-    const output = 'Expected output';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      assertion: javascriptMultilineStringAssertion,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-    });
-    expect(result).toMatchObject({
-      pass: true,
-      reason: 'Assertion passed',
-    });
-  });
-
-  it('should pass when the multiline javascript assertion fails', async () => {
-    const output = 'Not the expected output';
-
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      assertion: javascriptMultilineStringAssertion,
-      test: {} as AtomicTestCase,
-      providerResponse: { output },
-      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-    });
-    expect(result).toMatchObject({
-      pass: false,
-      reason: 'Assertion failed',
     });
   });
 
@@ -2370,390 +2044,6 @@ describe('runAssertion', () => {
     expect(result).toMatchObject({
       pass: false,
       reason: 'Levenshtein distance 8 is greater than threshold 5',
-    });
-  });
-
-  it.each([
-    [
-      'boolean',
-      jest.fn((output: string) => output === 'Expected output'),
-      true,
-      'Assertion passed',
-    ],
-    ['number', jest.fn((output: string) => output.length), true, 'Assertion passed'],
-    [
-      'GradingResult',
-      jest.fn((output: string) => ({ pass: true, score: 1, reason: 'Custom reason' })),
-      true,
-      'Custom reason',
-    ],
-    [
-      'boolean',
-      jest.fn((output: string) => output !== 'Expected output'),
-      false,
-      'Custom function returned false',
-    ],
-    ['number', jest.fn((output: string) => 0), false, 'Custom function returned false'],
-    [
-      'GradingResult',
-      jest.fn((output: string) => ({ pass: false, score: 0.1, reason: 'Custom reason' })),
-      false,
-      'Custom reason',
-    ],
-    [
-      'boolean Promise',
-      jest.fn((output: string) => Promise.resolve(true)),
-      true,
-      'Assertion passed',
-    ],
-  ])(
-    'should pass when the file:// assertion with .js file returns a %s',
-    async (type, mockFn, expectedPass, expectedReason) => {
-      const output = 'Expected output';
-
-      jest.doMock(path.resolve('/path/to/assert.js'), () => mockFn, { virtual: true });
-
-      const fileAssertion: Assertion = {
-        type: 'javascript',
-        value: 'file:///path/to/assert.js',
-      };
-
-      const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-      const providerResponse = { output };
-      const result: GradingResult = await runAssertion({
-        prompt: 'Some prompt',
-        provider,
-        assertion: fileAssertion,
-        test: {} as AtomicTestCase,
-        providerResponse,
-      });
-
-      expect(mockFn).toHaveBeenCalledWith('Expected output', {
-        prompt: 'Some prompt',
-        vars: {},
-        test: {},
-        provider,
-        providerResponse,
-      });
-      expect(result).toMatchObject({
-        pass: expectedPass,
-        reason: expect.stringContaining(expectedReason),
-      });
-    },
-  );
-
-  it.each([
-    [
-      'boolean',
-      jest.fn((output: string) => output === 'Expected output'),
-      true,
-      'Assertion passed',
-    ],
-    ['number', jest.fn((output: string) => output.length), true, 'Assertion passed'],
-    [
-      'GradingResult',
-      jest.fn((output: string) => ({ pass: true, score: 1, reason: 'Custom reason' })),
-      true,
-      'Custom reason',
-    ],
-    [
-      'boolean',
-      jest.fn((output: string) => output !== 'Expected output'),
-      false,
-      'Custom function returned false',
-    ],
-    ['number', jest.fn((output: string) => 0), false, 'Custom function returned false'],
-    [
-      'GradingResult',
-      jest.fn((output: string) => ({ pass: false, score: 0.1, reason: 'Custom reason' })),
-      false,
-      'Custom reason',
-    ],
-    [
-      'boolean Promise',
-      jest.fn((output: string) => Promise.resolve(true)),
-      true,
-      'Assertion passed',
-    ],
-  ])(
-    'should pass when assertion is a package path',
-    async (type, mockFn, expectedPass, expectedReason) => {
-      const output = 'Expected output';
-
-      const require = createRequire('');
-      jest.spyOn(require, 'resolve').mockReturnValueOnce('/node_modules/@promptfoo/fake/index.js');
-      jest.doMock(
-        path.resolve('/node_modules/@promptfoo/fake/index.js'),
-        () => {
-          return {
-            assertionFunction: mockFn,
-          };
-        },
-        { virtual: true },
-      );
-
-      const fileAssertion: Assertion = {
-        type: 'javascript',
-        value: 'package:@promptfoo/fake:assertionFunction',
-      };
-
-      const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-      const providerResponse = { output };
-      const result: GradingResult = await runAssertion({
-        prompt: 'Some prompt',
-        provider,
-        assertion: fileAssertion,
-        test: {} as AtomicTestCase,
-        providerResponse,
-      });
-
-      expect(mockFn).toHaveBeenCalledWith('Expected output', {
-        prompt: 'Some prompt',
-        vars: {},
-        test: {},
-        provider,
-        providerResponse,
-      });
-      expect(result).toMatchObject({
-        pass: expectedPass,
-        reason: expect.stringContaining(expectedReason),
-      });
-    },
-  );
-
-  it('should resolve js paths relative to the configuration file', async () => {
-    const output = 'Expected output';
-    const mockFn = jest.fn((output: string) => output === 'Expected output');
-
-    jest.doMock(path.resolve('/base/path/path/to/assert.js'), () => mockFn, { virtual: true });
-
-    const fileAssertion: Assertion = {
-      type: 'javascript',
-      value: 'file://./path/to/assert.js',
-    };
-
-    const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-    const providerResponse = { output };
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider,
-      assertion: fileAssertion,
-      test: {} as AtomicTestCase,
-      providerResponse,
-    });
-
-    expect(mockFn).toHaveBeenCalledWith('Expected output', {
-      prompt: 'Some prompt',
-      vars: {},
-      test: {},
-      provider,
-      providerResponse,
-    });
-    expect(result).toMatchObject({
-      pass: true,
-      reason: 'Assertion passed',
-    });
-  });
-
-  it('should handle output strings with both single and double quotes correctly in python assertion', async () => {
-    const expectedPythonValue = '0.5';
-
-    jest.mocked(runPythonCode).mockResolvedValueOnce(expectedPythonValue);
-
-    const output =
-      'This is a string with "double quotes"\n and \'single quotes\' \n\n and some \n\t newlines.';
-
-    const pythonAssertion: Assertion = {
-      type: 'python',
-      value: expectedPythonValue,
-    };
-
-    const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-    const providerResponse = { output };
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt',
-      provider,
-      assertion: pythonAssertion,
-      test: {} as AtomicTestCase,
-      providerResponse,
-    });
-
-    expect(runPythonCode).toHaveBeenCalledTimes(1);
-    expect(runPythonCode).toHaveBeenCalledWith(expect.anything(), 'main', [
-      output,
-      { prompt: 'Some prompt', test: {}, vars: {}, provider, providerResponse },
-    ]);
-
-    expect(result).toMatchObject({
-      pass: true,
-      reason: 'Assertion passed',
-      score: Number(expectedPythonValue),
-    });
-  });
-
-  it.each([
-    ['boolean', false, 0, 'Python code returned false', false, undefined],
-    ['number', 0, 0, 'Python code returned false', false, undefined],
-    [
-      'GradingResult',
-      `{"pass": false, "score": 0, "reason": "Custom error"}`,
-      0,
-      'Custom error',
-      false,
-      undefined,
-    ],
-    ['boolean', true, 1, 'Assertion passed', true, undefined],
-    ['number', 1, 1, 'Assertion passed', true, undefined],
-    [
-      'GradingResult',
-      `{"pass": true, "score": 1, "reason": "Custom success"}`,
-      1,
-      'Custom success',
-      true,
-      undefined,
-    ],
-    [
-      'GradingResult',
-      // This score is less than the assertion threshold in the test
-      `{"pass": true, "score": 0.4, "reason": "Foo bar"}`,
-      0.4,
-      'Python score 0.4 is less than threshold 0.5',
-      false,
-      0.5,
-    ],
-  ])(
-    'should handle inline return type %s with return value: %p',
-    async (type, returnValue, expectedScore, expectedReason, expectedPass, threshold) => {
-      const output =
-        'This is a string with "double quotes"\n and \'single quotes\' \n\n and some \n\t newlines.';
-
-      let resolvedValue;
-      if (type === 'GradingResult') {
-        resolvedValue = JSON.parse(returnValue as string);
-      } else {
-        resolvedValue = returnValue;
-      }
-
-      const pythonAssertion: Assertion = {
-        type: 'python',
-        value: returnValue.toString(),
-        threshold,
-      };
-
-      jest.mocked(runPythonCode).mockResolvedValueOnce(resolvedValue);
-
-      const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-      const providerResponse = { output };
-      const result: GradingResult = await runAssertion({
-        prompt: 'Some prompt',
-        provider,
-        assertion: pythonAssertion,
-        test: {} as AtomicTestCase,
-        providerResponse,
-      });
-
-      expect(runPythonCode).toHaveBeenCalledTimes(1);
-      expect(runPythonCode).toHaveBeenCalledWith(expect.anything(), 'main', [
-        output,
-        { prompt: 'Some prompt', test: {}, vars: {}, provider, providerResponse },
-      ]);
-
-      expect(result).toMatchObject({
-        pass: expectedPass,
-        reason: expect.stringMatching(expectedReason),
-        score: expectedScore,
-      });
-    },
-  );
-
-  it.each([
-    ['boolean', 'True', true, 'Assertion passed'],
-    ['number', '0.5', true, 'Assertion passed'],
-    ['boolean', true, true, 'Assertion passed'],
-    ['number', 0.5, true, 'Assertion passed'],
-    [
-      'GradingResult',
-      '{"pass": true, "score": 1, "reason": "Custom reason"}',
-      true,
-      'Custom reason',
-    ],
-    ['boolean', 'False', false, 'Python code returned false'],
-    ['number', '0', false, 'Python code returned false'],
-    [
-      'GradingResult',
-      '{"pass": false, "score": 0, "reason": "Custom reason"}',
-      false,
-      'Custom reason',
-    ],
-  ])(
-    'should handle when the file:// assertion with .py file returns a %s',
-    async (type, pythonOutput, expectedPass, expectedReason) => {
-      const output = 'Expected output';
-      jest.mocked(runPython).mockResolvedValueOnce(pythonOutput as string | object);
-
-      const fileAssertion: Assertion = {
-        type: 'python',
-        value: 'file:///path/to/assert.py',
-      };
-
-      const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-      const providerResponse = { output };
-      const result: GradingResult = await runAssertion({
-        prompt: 'Some prompt that includes "double quotes" and \'single quotes\'',
-        provider,
-        assertion: fileAssertion,
-        test: {} as AtomicTestCase,
-        providerResponse,
-      });
-
-      expect(runPython).toHaveBeenCalledWith(path.resolve('/path/to/assert.py'), 'get_assert', [
-        output,
-        {
-          prompt: 'Some prompt that includes "double quotes" and \'single quotes\'',
-          vars: {},
-          test: {},
-          provider,
-          providerResponse,
-        },
-      ]);
-
-      expect(result).toMatchObject({
-        pass: expectedPass,
-        reason: expect.stringContaining(expectedReason),
-      });
-      expect(runPython).toHaveBeenCalledTimes(1);
-    },
-  );
-
-  it('should handle when python file assertions throw an error', async () => {
-    const output = 'Expected output';
-    jest
-      .mocked(runPython)
-      .mockRejectedValue(
-        new Error('The Python script `call_api` function must return a dict with an `output`'),
-      );
-    const fileAssertion: Assertion = {
-      type: 'python',
-      value: 'file:///path/to/assert.py',
-    };
-    const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-    const providerResponse = { output };
-    const result: GradingResult = await runAssertion({
-      prompt: 'Some prompt that includes "double quotes" and \'single quotes\'',
-      provider,
-      assertion: fileAssertion,
-      test: {} as AtomicTestCase,
-      providerResponse,
-    });
-    expect(runPython).toHaveBeenCalledTimes(1);
-    expect(result).toEqual({
-      assertion: {
-        type: 'python',
-        value: 'file:///path/to/assert.py',
-      },
-      pass: false,
-      reason: 'The Python script `call_api` function must return a dict with an `output`',
-      score: 0,
     });
   });
 
@@ -3937,151 +3227,6 @@ describe('runAssertion', () => {
 
       expect(fs.readFileSync).toHaveBeenCalledWith('/base/path/schema.json', 'utf8');
       expect(result.pass).toBe(true);
-    });
-
-    it('should handle JavaScript file reference with function name', async () => {
-      const assertion: Assertion = {
-        type: 'javascript',
-        value: 'file:///path/to/assert.js:customFunction',
-      };
-
-      const mockFn = jest.fn((output: string) => true);
-      jest.doMock(
-        path.resolve('/path/to/assert.js'),
-        () => ({
-          customFunction: mockFn,
-        }),
-        { virtual: true },
-      );
-
-      const output = 'Expected output';
-      const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-      const providerResponse = { output };
-
-      const result: GradingResult = await runAssertion({
-        prompt: 'Some prompt',
-        provider,
-        assertion,
-        test: {} as AtomicTestCase,
-        providerResponse,
-      });
-
-      expect(mockFn).toHaveBeenCalledWith('Expected output', {
-        prompt: 'Some prompt',
-        vars: {},
-        test: {},
-        provider,
-        providerResponse,
-      });
-      expect(result).toMatchObject({
-        pass: true,
-        reason: 'Assertion passed',
-      });
-    });
-
-    it('should handle Python file reference with function name', async () => {
-      const assertion: Assertion = {
-        type: 'python',
-        value: 'file:///path/to/assert.py:custom_function',
-      };
-
-      const mockOutput = true;
-      jest.mocked(runPython).mockResolvedValueOnce(mockOutput);
-
-      const output = 'Expected output';
-      const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-      const providerResponse = { output };
-
-      const result: GradingResult = await runAssertion({
-        prompt: 'Some prompt',
-        provider,
-        assertion,
-        test: {} as AtomicTestCase,
-        providerResponse,
-      });
-
-      expect(runPython).toHaveBeenCalledWith(
-        path.resolve('/path/to/assert.py'),
-        'custom_function',
-        [
-          output,
-          {
-            prompt: 'Some prompt',
-            vars: {},
-            test: {},
-            provider,
-            providerResponse,
-          },
-        ],
-      );
-      expect(result).toMatchObject({
-        pass: true,
-        reason: 'Assertion passed',
-      });
-    });
-
-    it('should ignore function name in file path for non-JS/Python files', async () => {
-      const assertion: Assertion = {
-        type: 'equals',
-        value: 'file:///path/to/text.txt:ignored',
-      };
-
-      const expectedContent = 'Expected output';
-      jest.mocked(fs.readFileSync).mockReturnValue(expectedContent);
-      jest.mocked(path.resolve).mockReturnValue('/base/path/path/to/text.txt:ignored');
-      jest.mocked(path.extname).mockReturnValue('.txt');
-
-      const result: GradingResult = await runAssertion({
-        prompt: 'Some prompt',
-        provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
-        assertion,
-        test: {} as AtomicTestCase,
-        providerResponse: { output: 'Expected output' },
-      });
-
-      expect(fs.readFileSync).toHaveBeenCalledWith('/base/path/path/to/text.txt:ignored', 'utf8');
-      expect(result.pass).toBe(true);
-    });
-
-    it('should use default function name for Python when none specified', async () => {
-      const assertion: Assertion = {
-        type: 'python',
-        value: 'file:///path/to/assert.py',
-      };
-
-      const mockOutput = true;
-      jest.mocked(runPython).mockResolvedValueOnce(mockOutput);
-
-      const output = 'Expected output';
-      const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
-      const providerResponse = { output };
-
-      const result: GradingResult = await runAssertion({
-        prompt: 'Some prompt',
-        provider,
-        assertion,
-        test: {} as AtomicTestCase,
-        providerResponse,
-      });
-
-      expect(runPython).toHaveBeenCalledWith(
-        path.resolve('/path/to/assert.py'),
-        'get_assert', // Default function name
-        [
-          output,
-          {
-            prompt: 'Some prompt',
-            vars: {},
-            test: {},
-            provider,
-            providerResponse,
-          },
-        ],
-      );
-      expect(result).toMatchObject({
-        pass: true,
-        reason: 'Assertion passed',
-      });
     });
   });
 });
